Add support for `es6` module build
Use `babel-cli` instead of `babel-core` so we can call `babel` from npm scripts.

Delete `.babelrc` because it apparently overrides all CLI commands, which means we can't adapt which presets we use based on what the user is trying to do.

For example, to output es6 modules, we need to use `react` preset, but not `es2015`.

So, now both `webpack` and the npm scripts pass their own babel configs.