Add webpack config for fonts and images
Add loader configs for font and image formats to the `common` config.

In the original webpack config, it would freak out when you tried to load something like Bootstrap because all the font/svg files weren't accounted for.

This new config should enable the use of a CSS framework (like bootstrap) right out of the box since webpack will handle the font and image files automatically.